+++ bluetooth_low_energy_platform_interface/lib/src/bluetooth_low_energy_manager.dart
@@ -1,14 +1,13 @@
 import 'bluetooth_low_energy_event_args.dart';
 import 'bluetooth_low_energy_state.dart';
+import 'logger_controller.dart';
+import 'set_up.dart';
 
 /// The abstract base class that manages central and peripheral objects.
-abstract class BluetoothLowEnergyManager {
+abstract class BluetoothLowEnergyManager implements SetUp, LoggerController {
   /// The current state of the manager.
   BluetoothLowEnergyState get state;
 
   /// Tells the manager’s state updated.
   Stream<BluetoothLowEnergyStateChangedEventArgs> get stateChanged;
-
-  /// Sets up this bluetooth low energy manager.
-  Future<void> setUp();
 }