+++ bluetooth_low_energy_platform_interface/lib/src/my_central_manager.dart
@@ -1,14 +1,15 @@
+import 'package:logging/logging.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 import 'central_manager.dart';
+import 'my_logger_controller.dart';
+import 'my_set_up.dart';
 
 /// Platform-specific implementations should implement this class to support
 /// [CentralManager].
 abstract class MyCentralManager extends PlatformInterface
+    with MySetUp, MyLoggerController
     implements CentralManager {
-  /// Constructs a [MyCentralManager].
-  MyCentralManager() : super(token: _token);
-
   static final Object _token = Object();
 
   static MyCentralManager? _instance;
@@ -31,4 +32,12 @@ abstract class MyCentralManager extends PlatformInterface
     PlatformInterface.verifyToken(instance, _token);
     _instance = instance;
   }
+
+  @override
+  final Logger logger;
+
+  /// Constructs a [MyCentralManager].
+  MyCentralManager()
+      : logger = Logger('MyCentralManager'),
+        super(token: _token);
 }