+++ bluetooth_low_energy_platform_interface/lib/src/my_peripheral_manager.dart
@@ -1,14 +1,15 @@
+import 'package:logging/logging.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
+import 'my_logger_controller.dart';
+import 'my_set_up.dart';
 import 'peripheral_manager.dart';
 
 /// Platform-specific implementations should implement this class to support
 /// [PeripheralManager].
 abstract class MyPeripheralManager extends PlatformInterface
+    with MySetUp, MyLoggerController
     implements PeripheralManager {
-  /// Constructs a [MyPeripheralManager].
-  MyPeripheralManager() : super(token: _token);
-
   static final Object _token = Object();
 
   static MyPeripheralManager? _instance;
@@ -31,4 +32,12 @@ abstract class MyPeripheralManager extends PlatformInterface
     PlatformInterface.verifyToken(instance, _token);
     _instance = instance;
   }
+
+  @override
+  final Logger logger;
+
+  /// Constructs a [MyPeripheralManager].
+  MyPeripheralManager()
+      : logger = Logger('MyPeripheralManager'),
+        super(token: _token);
 }