+++ bluetooth_low_energy_platform_interface/lib/src/my_set_up.dart
@@ -1,0 +1,9 @@
+import 'package:meta/meta.dart';
+
+import 'set_up.dart';
+
+mixin MySetUp implements SetUp {
+  @override
+  @mustCallSuper
+  Future<void> setUp() async {}
+}