+++ bluetooth_low_energy_platform_interface/lib/src/set_up.dart
@@ -1,0 +1,5 @@
+/// Sets up an instance.
+abstract class SetUp {
+  /// Sets up the instance.
+  Future<void> setUp();
+}